Added `CloudEventImmutable` implementation
